requirements.txt
# Flask dan ekstensi
Flask==2.2.3
Flask-Session==0.4.0
Werkzeug==2.2.3

# Database dan manipulasi data
pandas==1.5.3
numpy==1.24.2

# Deep Learning dan NLP
torch==2.0.0
transformers==4.27.2
nltk==3.8.1

# Visualisasi
matplotlib==3.7.1
seaborn==0.12.2

# Generasi PDF
reportlab==3.6.12

# API AI
google-generativeai==0.3.0

# Utilitas
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
requests==2.28.2
tqdm==4.65.0
python-dateutil==2.8.2
pytz==2023.3